test(qa-channel): split direct action coverage
