+++ extensions/qa-channel/src/channel.test.ts
@@ -798,57 +798,6 @@ describe("qa-channel plugin", () => {
     }
   });
 
-  it("prefers canonical targets over conflicting direct-adapter aliases", async () => {
-    installQaChannelTestRegistry();
-    const state = createQaBusState();
-    const bus = await startQaBusServer({ state });
-
-    try {
-      const cfg = createQaChannelConfig({ baseUrl: bus.baseUrl });
-      const handleAction = requireQaActionHandler();
-      const threadResult = await handleAction({
-        channel: "qa-channel",
-        action: "thread-create",
-        cfg,
-        accountId: "default",
-        params: {
-          target: "channel:canonical-room",
-          to: "channel:legacy-to-room",
-          channelId: "legacy-channel-id-room",
-          threadName: "Canonical target thread",
-        },
-      });
-      const threadPayload = extractToolPayload(threadResult) as {
-        thread: { id: string; conversationId: string };
-        target: string;
-      };
-
-      expect(threadPayload.thread.conversationId).toBe("canonical-room");
-      expect(threadPayload.target).toBe(`thread:canonical-room/${threadPayload.thread.id}`);
-
-      const replyResult = await handleAction({
-        channel: "qa-channel",
-        action: "thread-reply",
-        cfg,
-        accountId: "default",
-        params: {
-          target: threadPayload.target,
-          channelId: "legacy-reply-room",
-          message: "canonical target reply",
-        },
-      });
-      expect(extractToolPayload(replyResult)).toMatchObject({
-        message: {
-          conversation: { id: "canonical-room", kind: "channel" },
-          text: "canonical target reply",
-          threadId: threadPayload.thread.id,
-        },
-      });
-    } finally {
-      await bus.stop();
-    }
-  });
-
   it("rejects thread replies outside the owning account and conversation", async () => {
     installQaChannelTestRegistry();
     const state = createQaBusState();
@@ -898,73 +847,6 @@ describe("qa-channel plugin", () => {
       });
       const payload = extractToolPayload(result) as { message: { threadId: string } };
       expect(payload.message.threadId).toBe(thread.id);
-    } finally {
-      await bus.stop();
-    }
-  });
-
-  it("keeps the shipped thread and edit aliases for direct API callers", async () => {
-    installQaChannelTestRegistry();
-    const state = createQaBusState();
-    const bus = await startQaBusServer({ state });
-
-    try {
-      const cfg = createQaChannelConfig({ baseUrl: bus.baseUrl });
-      const handleAction = requireQaActionHandler();
-      const threadResult = await handleAction({
-        channel: "qa-channel",
-        action: "thread-create",
-        cfg,
-        accountId: "default",
-        params: {
-          channelId: "qa-room",
-          title: "Legacy thread",
-        },
-      });
-      const threadPayload = extractToolPayload(threadResult) as {
-        thread: { id: string; title: string };
-        target: string;
-      };
-      expect(threadPayload.thread.title).toBe("Legacy thread");
-
-      const replyResult = await handleAction({
-        channel: "qa-channel",
-        action: "thread-reply",
-        cfg,
-        accountId: "default",
-        params: {
-          channelId: "qa-room",
-          threadId: threadPayload.thread.id,
-          text: "legacy reply",
-        },
-      });
-      expect(extractToolPayload(replyResult)).toMatchObject({
-        message: {
-          text: "legacy reply",
-          threadId: threadPayload.thread.id,
-        },
-      });
-
-      const outbound = state.addOutboundMessage({
-        to: threadPayload.target,
-        text: "before legacy edit",
-        threadId: threadPayload.thread.id,
-      });
-      const editResult = await handleAction({
-        channel: "qa-channel",
-        action: "edit",
-        cfg,
-        accountId: "default",
-        params: {
-          to: threadPayload.target,
-          messageId: outbound.id,
-          text: "legacy edit",
-        },
-      });
-      expect(extractToolPayload(editResult)).toMatchObject({
-        message: { text: "legacy edit" },
-      });
-      expect(state.readMessage({ messageId: outbound.id }).text).toBe("legacy edit");
     } finally {
       await bus.stop();
     }
